Calculator protocol and README updates for logger and calculator
