+++ src/calculator/pyproject.toml
@@ -1,10 +1,8 @@
 [project]
 name = "calculator"
 version = "0.1.0"
-description = "Calculator module"
+description = "Calculator Protocol"
+readme = "README.md"
 requires-python = ">=3.13"
 dependencies = []
-
-[build-system]
-requires = ["setuptools"]
-build-backend = "setuptools.build_meta"
+
