+++ src/calculator_impl/pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "calculator-impl"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "calculator",
+]
+
+[tool.uv.sources]
+notifier = { path = "../calculator", editable = true }
