+++ src/calculator_impl/src/calculator_impl/__init__.py
@@ -1,0 +1,9 @@
+"""Module for a Calculator implementation."""
+
+from typing import Optional
+import calculator
+
+from . import _impl
+
+
+calculator.get_calculator = lambda: _impl.Calculator()
